# docs/swagger.yaml
basePath: /
definitions:
...
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.Camion:
    properties:
      camion_id:
        type: integer
      color_disponibilidad:
        type: string
      created_at:
        type: string
      disponibilidad_id:
        type: integer
      es_rentado:
        type: boolean
      modelo:
        type: string
      nombre_disponibilidad:
        type: string
      placa:
        type: string
      tipo_camion_id:
        type: integer
      updated_at:
        type: string
    type: object
  entities.CamionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Camion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.CreateCamionRequest:
    properties:
      disponibilidad_id:
        type: integer
      es_rentado:
        type: boolean
      modelo:
        type: string
      placa:
        type: string
      tipo_camion_id:
        type: integer
    required:
    - disponibilidad_id
    - modelo
    - placa
    - tipo_camion_id
    type: object
...
  entities.CreatePuntoRecoleccionRequest:
    properties:
      cp:
        type: string
      lat:
        type: number
      lon:
        type: number
      punto_id:
        type: integer
      ruta_id:
        type: integer
    type: object
...
  entities.EstadoCamion:
    properties:
      camion_id:
        type: integer
      estado:
        description: nullable
        type: string
      estado_id:
        type: integer
      observaciones:
        description: nullable
        type: string
      timestamp:
        description: nullable
        type: string
    type: object
  entities.EstadoCamionListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.EstadoCamion'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.EstadoCamionMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.EstadoCamionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.EstadoCamion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.PuntoRecoleccion:
    properties:
      cp:
        type: string
      created_at:
        type: string
      eliminado:
        type: boolean
      lat:
        type: number
      lon:
        type: number
      punto_id:
        type: integer
      ruta_id:
        type: integer
    type: object
  entities.PuntoRecoleccionListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.PuntoRecoleccion'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.PuntoRecoleccionMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.PuntoRecoleccionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.PuntoRecoleccion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.RellenoSanitario:
    properties:
      capacidad_toneladas:
        type: number
      direccion:
        type: string
      eliminado:
        type: boolean
      es_rentado:
        type: boolean
      nombre:
        type: string
      relleno_id:
        type: integer
    type: object
  entities.RellenoSanitarioListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.RellenoSanitario'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.UpdateCamionRequest:
    properties:
      disponibilidad_id:
        type: integer
      es_rentado:
        type: boolean
      modelo:
        type: string
      placa:
        type: string
      tipo_camion_id:
        type: integer
    type: object
...
  entities.UpdateEstadoCamionRequest:
    properties:
      camion_id:
        type: integer
      estado:
        type: string
      estado_id:
        type: integer
      observaciones:
        type: string
      timestamp:
        type: string
    type: object
...
  entities.UpdatePuntoRecoleccionRequest:
    properties:
      cp:
        type: string
      lat:
        type: number
      lon:
        type: number
      punto_id:
        type: integer
      ruta_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/camion/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar camiones
      tags:
      - Camion
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateCamionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.CamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear camión
      tags:
      - Camion
  /api/camion/{id}:
    delete:
      parameters:
      - description: ID camión
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar camión
      tags:
      - Camion
    get:
      parameters:
      - description: ID camión
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener camión por ID
      tags:
      - Camion
    put:
      consumes:
      - application/json
      parameters:
      - description: ID camión
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateCamionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar camión
      tags:
      - Camion
...
  /api/camion/placa/{placa}:
    get:
      parameters:
      - description: Placa
        in: path
        name: placa
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar camión por placa
      tags:
      - Camion
...
  /api/puntos-recoleccion/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PuntoRecoleccionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar puntos de recolección
      tags:
      - PuntoRecoleccion
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreatePuntoRecoleccionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.PuntoRecoleccionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear punto de recolección
      tags:
      - PuntoRecoleccion
  /api/puntos-recoleccion/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PuntoRecoleccionMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar punto de recolección
      tags:
      - PuntoRecoleccion
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PuntoRecoleccionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Punto de recolección por ID
      tags:
      - PuntoRecoleccion
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdatePuntoRecoleccionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PuntoRecoleccionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar punto de recolección
      tags:
      - PuntoRecoleccion
  /api/puntos-recoleccion/ruta/{rutaId}:
    get:
      parameters:
      - description: ID ruta
        in: path
        name: rutaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PuntoRecoleccionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Puntos de recolección por ruta
      tags:
      - PuntoRecoleccion
...
  /api/relleno-sanitario/exists/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RellenoSanitarioListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verificar existencia relleno sanitario
      tags:
      - RellenoSanitario
...
  /api/rutas/{id}:
    delete:
      parameters:
      - description: ID ruta
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar ruta
      tags:
      - Ruta
    get:
      parameters:
      - description: ID ruta
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener ruta por ID
      tags:
      - Ruta
    put:
      consumes:
      - application/json
      parameters:
      - description: ID ruta
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateEstadoCamionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar ruta
      tags:
      - Ruta
...
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"